requirements.txt
Flask==0.12
pycryptodome==3.15.0